+++ src/pages/writing/BasicSentence.jsx
@@ -3,7 +3,6 @@ import { Link } from 'react-router-dom';
 import SEOHead from '../../components/SEOHead';
 import { useLanguage } from '../../contexts/LanguageContext';
 import useAOS from '../../hooks/useAOS';
-import SubNav from '../../components/SubNav';
 import PageLayout from '../../components/PageLayout';
 
 const sentenceStructures = [
@@ -192,8 +191,7 @@ function BasicSentence() {
         </div>
       </section>
 
-      <SubNav category="writing" />
-      <PageLayout sections={sections}>
+      <PageLayout sections={sections} category="writing">
         <div className="content-page">
 
         {/* Sentence Structures */}